+++ lib/src/mutation_state.dart
@@ -2,7 +2,10 @@ import 'package:meta/meta.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:stack_trace/stack_trace.dart';
 
-/// AsyncValue-like data class, but includes a ref and callback
+typedef StateSetter<Result, Param> = void Function(
+    MutationState<Result, Param> state);
+
+/// AsyncValue-like data class, but includes a setState and callback
 /// strictly for use in providers
 /// `_fn` is the handler for whatever parameter gets passed in (a call to a notifier function, for example)
 /// Param is the type of the value that gets passed in by the front end
@@ -10,56 +13,65 @@ import 'package:stack_trace/stack_trace.dart';
 /// (the type of the `value`)
 @immutable
 abstract class MutationState<Result, Param> {
-  const MutationState._(this._ref, this._fn);
+  const MutationState(this._setState, this._fn);
 
-  /// Use to initialize your own mutation provider
-  ///
-  /// [ref] is the provider reference
-  ///
-  /// [fn] is the function that will be called when the mutation is called
-  /// with the parameter passed in by the front end
-  factory MutationState.create(
+  factory MutationState.fromRef(
     ProviderRef<MutationState<Result, Param>> ref,
     FutureOr<Result> Function(Param p) fn,
   ) =>
-      MutationState<Result, Param>._initial(ref, fn);
+      MutationState<Result, Param>.initial(
+        (state) => ref.state = state,
+        fn,
+      );
 
-  const factory MutationState._initial(
-    ProviderRef<MutationState<Result, Param>> ref,
+  /// Use to initialize your own mutation provider
+  ///
+  /// [setState] is the state setter that will be called when the mutation
+  /// is called with the parameter passed in by the front end
+  ///
+  /// [fn] is the function that will be called when the mutation is called
+  /// with the parameter passed in by the front end
+  const factory MutationState.create(
+    StateSetter<Result, Param> setState,
     FutureOr<Result> Function(Param p) fn,
-  ) = MutationInitial<Result, Param>._;
+  ) = MutationState<Result, Param>.initial;
 
-  const factory MutationState._error(
-    ProviderRef<MutationState<Result, Param>> ref,
+  const factory MutationState.initial(
+    StateSetter<Result, Param> setState,
+    FutureOr<Result> Function(Param p) fn,
+  ) = MutationInitial<Result, Param>;
+
+  const factory MutationState.error(
+    StateSetter<Result, Param> setState,
     FutureOr<Result> Function(Param p) fn,
     Object error, {
     required StackTrace stackTrace,
-  }) = MutationError<Result, Param>._;
+  }) = MutationError<Result, Param>;
 
-  const factory MutationState._data(
-    ProviderRef<MutationState<Result, Param>> ref,
+  const factory MutationState.data(
+    StateSetter<Result, Param> setState,
     FutureOr<Result> Function(Param p) fn,
     Result value,
-  ) = MutationData<Result, Param>._;
+  ) = MutationData<Result, Param>;
 
-  const factory MutationState._loading(
-    ProviderRef<MutationState<Result, Param>> ref,
+  const factory MutationState.loading(
+    StateSetter<Result, Param> setState,
     FutureOr<Result> Function(Param p) fn,
-  ) = MutationLoading<Result, Param>._;
+  ) = MutationLoading<Result, Param>;
 
-  static Future<MutationState<Result, P>> _guard<Result, P>(
-    ProviderRef<MutationState<Result, P>> ref,
+  static Future<MutationState<Result, P>> guard<Result, P>(
+    StateSetter<Result, P> setState,
     FutureOr<Result> Function(P p) cb,
     P parameter,
   ) async {
     try {
-      return MutationState<Result, P>._data(ref, cb, await cb(parameter));
+      return MutationState<Result, P>.data(setState, cb, await cb(parameter));
     } catch (e, s) {
-      return MutationState<Result, P>._error(ref, cb, e, stackTrace: s);
+      return MutationState<Result, P>.error(setState, cb, e, stackTrace: s);
     }
   }
 
-  final ProviderRef<MutationState<Result, Param>> _ref;
+  final StateSetter<Result, Param> _setState;
   bool get isLoading;
   bool get hasValue;
   Result? get value;
@@ -67,16 +79,15 @@ abstract class MutationState<Result, Param> {
   StackTrace? get stackTrace;
   final FutureOr<Result> Function(Param p) _fn;
 
-  void _setState(MutationState<Result, Param> state) {
-    _ref.state = state.copyWithPrevious(_ref.state);
+  void _mutateState(MutationState<Result, Param> newState) {
+    _setState(newState.copyWithPrevious(this));
   }
 
   Future<MutationState<Result, Param>> call(Param parameter) async {
-    final cb = this._fn;
-    _setState(MutationState<Result, Param>._loading(_ref, cb));
+    _mutateState(MutationState<Result, Param>.loading(_setState, _fn));
     final result =
-        await MutationState._guard<Result, Param>(_ref, cb, parameter);
-    _setState(result);
+        await MutationState.guard<Result, Param>(_setState, _fn, parameter);
+    _mutateState(result);
     return result;
   }
 
@@ -92,16 +103,18 @@ abstract class MutationState<Result, Param> {
 
   MutationState<Result, Param> unwrapPrevious() {
     return map(
-      initial: (i) => MutationInitial<Result, Param>._(_ref, _fn),
+      initial: (i) => MutationInitial<Result, Param>(_setState, _fn),
       data: (d) {
-        if (d.isLoading) return MutationLoading<Result, Param>._(_ref, _fn);
-        return MutationData._(_ref, _fn, d.value!);
+        if (d.isLoading) {
+          return MutationLoading<Result, Param>(_setState, _fn);
+        }
+        return MutationData(_setState, _fn, d.value!);
       },
       error: (e) {
-        if (e.isLoading) return MutationLoading._(_ref, _fn);
-        return MutationError._(_ref, _fn, e.error, stackTrace: e.stackTrace);
+        if (e.isLoading) return MutationLoading(_setState, _fn);
+        return MutationError(_setState, _fn, e.error, stackTrace: e.stackTrace);
       },
-      loading: (l) => MutationLoading<Result, Param>._(_ref, _fn),
+      loading: (l) => MutationLoading<Result, Param>(_setState, _fn),
     );
   }
 
@@ -144,7 +157,7 @@ abstract class MutationState<Result, Param> {
 }
 
 class MutationInitial<Result, Param> extends MutationState<Result, Param> {
-  const MutationInitial._(super._ref, super._fn) : super._();
+  const MutationInitial(super._setState, super._fn);
 
   @override
   Result? get value => null;
@@ -179,21 +192,20 @@ class MutationInitial<Result, Param> extends MutationState<Result, Param> {
 }
 
 class MutationLoading<Result, Param> extends MutationState<Result, Param> {
-  const MutationLoading._(super._ref, super._fn)
+  const MutationLoading(super._setState, super._fn)
       : hasValue = false,
         value = null,
         error = null,
-        stackTrace = null,
-        super._();
+        stackTrace = null;
 
-  const MutationLoading.__(
-    super._ref,
+  const MutationLoading._(
+    super._setState,
     super._fn, {
     required this.hasValue,
     required this.value,
     required this.error,
     required this.stackTrace,
-  }) : super._();
+  });
 
   @override
   bool get isLoading => true;
@@ -228,16 +240,16 @@ class MutationLoading<Result, Param> extends MutationState<Result, Param> {
     if (isRefresh) {
       return previous.map(
         initial: (_) => this,
-        data: (d) => MutationData.__(
-          _ref,
+        data: (d) => MutationData._(
+          _setState,
           _fn,
           d.value,
           isLoading: true,
           error: d.error,
           stackTrace: d.stackTrace,
         ),
-        error: (e) => MutationError.__(
-          _ref,
+        error: (e) => MutationError._(
+          _setState,
           _fn,
           e.error,
           isLoading: true,
@@ -250,16 +262,16 @@ class MutationLoading<Result, Param> extends MutationState<Result, Param> {
     } else {
       return previous.map(
         initial: (_) => this,
-        data: (e) => MutationLoading.__(
-          _ref,
+        data: (e) => MutationLoading._(
+          _setState,
           _fn,
           hasValue: true,
           value: e.valueOrNull,
           error: e.error,
           stackTrace: e.stackTrace,
         ),
-        error: (e) => MutationLoading.__(
-          _ref,
+        error: (e) => MutationLoading._(
+          _setState,
           _fn,
           hasValue: e.hasValue,
           value: e.valueOrNull,
@@ -273,20 +285,19 @@ class MutationLoading<Result, Param> extends MutationState<Result, Param> {
 }
 
 class MutationData<Result, Param> extends MutationState<Result, Param> {
-  const MutationData._(super._ref, super._fn, this.value)
+  const MutationData(super._setState, super._fn, this.value)
       : isLoading = false,
         error = null,
-        stackTrace = null,
-        super._();
+        stackTrace = null;
 
-  const MutationData.__(
-    super._ref,
+  const MutationData._(
+    super._setState,
     super._fn,
     this.value, {
     required this.isLoading,
     required this.error,
     required this.stackTrace,
-  }) : super._();
+  });
 
   @override
   final Result value;
@@ -320,23 +331,21 @@ class MutationData<Result, Param> extends MutationState<Result, Param> {
 }
 
 class MutationError<Result, Param> extends MutationState<Result, Param> {
-  const MutationError.__(
-    super._ref,
+  const MutationError(super._setState, super._fn, this.error,
+      {required this.stackTrace})
+      : isLoading = false,
+        hasValue = false,
+        _value = null;
+
+  const MutationError._(
+    super._setState,
     super._fn,
     this.error, {
     required this.stackTrace,
     required this.isLoading,
     required Result? value,
     required this.hasValue,
-  })  : _value = value,
-        super._();
-
-  const MutationError._(super._ref, super._fn, this.error,
-      {required this.stackTrace})
-      : isLoading = false,
-        hasValue = false,
-        _value = null,
-        super._();
+  }) : _value = value;
 
   @override
   final bool isLoading;
@@ -372,8 +381,8 @@ class MutationError<Result, Param> extends MutationState<Result, Param> {
   @override
   MutationState<Result, Param> copyWithPrevious(
       MutationState<Result, Param> previous) {
-    return MutationError.__(
-      _ref,
+    return MutationError._(
+      _setState,
       _fn,
       error,
       stackTrace: stackTrace,