+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog
+
+## 1.2.0
+
+- Exposed `MutationState` initial, data, loading, and error constructors for more generic use.
+
+- `MutationState` constructors no longer take a ref directly, see below.
+
+- Added `MutationState.fromRef` to replace the previous `MutationState.create` api
 
 ## 1.1.0